+++ src/RadixDlt.NetworkGateway.PostgresIntegration/CommonDbContext.cs
@@ -66,6 +66,7 @@ using Microsoft.EntityFrameworkCore;
 using RadixDlt.NetworkGateway.Abstractions;
 using RadixDlt.NetworkGateway.Abstractions.Model;
 using RadixDlt.NetworkGateway.Abstractions.Numerics;
+using RadixDlt.NetworkGateway.Abstractions.StandardMetadata;
 using RadixDlt.NetworkGateway.PostgresIntegration.Models;
 using RadixDlt.NetworkGateway.PostgresIntegration.ValueConverters;
 
@@ -78,7 +79,7 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration;
 /// </summary>
 internal abstract class CommonDbContext : DbContext
 {
-    private const string DiscriminatorColumnName = "discriminator";
+    internal const string DiscriminatorColumnName = "discriminator";
 
     /// <summary>
     /// Gets LedgerTransactions.
@@ -171,6 +172,10 @@ internal abstract class CommonDbContext : DbContext
 
     public DbSet<AccountAuthorizedDepositorAggregateHistory> AccountAuthorizedDepositorAggregateHistory => Set<AccountAuthorizedDepositorAggregateHistory>();
 
+    public DbSet<UnverifiedStandardMetadataAggregateHistory> UnverifiedStandardMetadataAggregateHistory => Set<UnverifiedStandardMetadataAggregateHistory>();
+
+    public DbSet<UnverifiedStandardMetadataEntryHistory> UnverifiedStandardMetadataEntryHistory => Set<UnverifiedStandardMetadataEntryHistory>();
+
     public CommonDbContext(DbContextOptions options)
         : base(options)
     {
@@ -201,6 +206,7 @@ internal abstract class CommonDbContext : DbContext
         modelBuilder.HasPostgresEnum<SborTypeKind>();
         modelBuilder.HasPostgresEnum<StateType>();
         modelBuilder.HasPostgresEnum<AuthorizedDepositorBadgeType>();
+        modelBuilder.HasPostgresEnum<StandardMetadataKey>();
 
         HookupTransactions(modelBuilder);
         HookupPendingTransactions(modelBuilder);
@@ -612,5 +618,23 @@ internal abstract class CommonDbContext : DbContext
         modelBuilder
             .Entity<ValidatorCumulativeEmissionHistory>()
             .HasIndex(e => new { e.ValidatorEntityId, e.EpochNumber });
+
+        modelBuilder
+            .Entity<UnverifiedStandardMetadataAggregateHistory>()
+            .HasIndex(e => new { e.EntityId, e.FromStateVersion });
+
+        modelBuilder
+            .Entity<UnverifiedStandardMetadataEntryHistory>()
+            .HasDiscriminator(e => e.Discriminator)
+            .HasValue<DappAccountTypeUnverifiedStandardMetadataEntryHistory>(StandardMetadataKey.DappAccountType)
+            .HasValue<DappDefinitionUnverifiedStandardMetadataEntryHistory>(StandardMetadataKey.DappDefinition)
+            .HasValue<DappDefinitionsUnverifiedStandardMetadataEntryHistory>(StandardMetadataKey.DappDefinitions)
+            .HasValue<DappClaimedEntitiesUnverifiedStandardMetadataEntryHistory>(StandardMetadataKey.DappClaimedEntities)
+            .HasValue<DappClaimedWebsitesUnverifiedStandardMetadataEntryHistory>(StandardMetadataKey.DappClaimedWebsites)
+            .HasValue<DappAccountLockerUnverifiedStandardMetadataEntryHistory>(StandardMetadataKey.DappAccountLocker);
+
+        modelBuilder
+            .Entity<UnverifiedStandardMetadataEntryHistory>()
+            .HasIndex(e => new { e.EntityId, e.Discriminator, e.FromStateVersion });
     }
 }